requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pytz>=2023.3
asgiref>=3.7.0
sqlparse>=0.5.0
django-mathfilters>=1.0.0
geopy>=2.4.0
openpyxl>=3.1.0
django-compressor>=4.5.1
psutil>=6.0.0
google-genai>=0.6.0
python-dotenv>=1.0.1
